Add mermaid source in Home MVVM
